Extracted state and handle non json value messages
