+++ src/HotTub.h
@@ -5,6 +5,7 @@
 #include "Commands.h"
 #include "Temperatures.h"
 #include "SendReceive.h"
+#include "State.h"
 
 #include <Syslog.h>
 
@@ -32,25 +33,6 @@
 #define PUMP_END 5
 #define PUMP_ERROR 6
 #define PUMP_FLASH 7
-
-struct CurrentState
-{
-  int pumpState;         //Pump's current state
-  int temperature;       //Pump's current actual water temperature
-  int targetTemperature; //Pump's current water temperature
-  bool temperatureIsCelsius;
-
-  bool flashing;                 //whether the display is blinking or not
-  unsigned int autoRestartCount; //the number of times autorestart has started the pump;
-
-  int errorCode;
-};
-
-struct TargetState
-{
-  int pumpState;         //State we want the pump to be in
-  int targetTemperature; //Target temperature we want to set the pump to
-};
 
 #define TM_NORMAL 0               //temperatures are current temperatures
 #define TM_TEMP_BUTTON_DETECTED 1 //ignore temperatures until we know if it's flashing
